// src/payload-types.ts
export interface BillingPlan {
  id: string;
  name: string;
  planId: string;
  price: number;
  price_formatted: string;
  interviewCount?: number | null;
  interviewDuration: number;
  freeTrialPeriod?: number | null;
  description?: string | null;
  features?:
    | {
        feature?: string | null;
        id?: string | null;
      }[]
    | null;
  buy_now_url: string;
  planDuration?: number | null;
  expiredIn?: number | null;
  updatedAt: string;
  createdAt: string;
}